pyproject.toml
[tool.poetry]
name = "darnitsacashbot"
version = "0.1.0"
description = "End-to-end automation platform for DarnitsaCashBot campaign."
authors = ["Darnitsa Automation <eng@darnitsa.ua>"]
readme = "README.md"
packages = [
  { include = "apps", from = "." },
  { include = "libs", from = "." },
  { include = "services", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.121.1"
uvicorn = { extras = ["standard"], version = "^0.33.0" }
aiogram = "^3.13.1"
aioboto3 = "^15.5.0"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
sqlalchemy = "^2.0.44"
alembic = "^1.14.0"
asyncpg = "^0.31.0"
httpx = "^0.28.1"
redis = "^7.1.0"
python-multipart = "^0.0.20"
tenacity = "^9.0.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
cryptography = "^46.0.3"
pendulum = "^3.0.0"
numpy = "^2.1.0"
opencv-python-headless = "^4.12.0.88"
Pillow = "^11.0.0"
pytesseract = "^0.3.10"
python-Levenshtein = "^0.27.0"
Unidecode = "^1.3.8"
qreader = "^0.4.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.7.0"
mypy = "^1.11.0"
types-redis = "^4.6.0"
types-pytz = "^2024.1"
fakeredis = "^2.25.0"
aiosqlite = "^0.21.0"

[tool.poetry.scripts]
telegram-bot = "apps.telegram_bot.main:run"
api-gateway = "apps.api_gateway.main:run"
bonus-service = "services.bonus_service.main:run_worker"
ocr-worker = "services.ocr_worker.worker:run_worker"
rules-engine = "services.rules_engine.service:run_worker"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
extend-select = ["I"]
target-version = "py311"

[tool.ruff.lint.isort]
known-first-party = ["apps", "libs", "services"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra -q --maxfail=1"

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
ignore_missing_imports = true

